requirements.txt
fastapi
uvicorn
aiofiles
python-dotenv
openai>=1.53.0
pinecone>=6.0.2
PyPDF2
pandas
numpy
ragas
langchain
datasets
bert-score
transformers
torch
nest-asyncio
loguru
requests